Arrumada listagem de marca em produto Arrumadas telas de atualizar para deixar os campos de letras maiúsculos ao salvar no BD Adicionado método de listar por id ou nome em fornecedor para importação por excel

+++ app/src/main/java/com/vandamodaintima/jfpsb/contador/model/dao/DAOProduto.java
@@ -222,7 +222,7 @@ public class DAOProduto implements DAO<Produto> {
 
                 cursorCodigoBarraFornecedor.close();
 
-                produto.setMarca(daoMarca.listarPorId(cursor.getInt(cursor.getColumnIndexOrThrow("marca"))));
+                produto.setMarca(daoMarca.listarPorId(cursor.getLong(cursor.getColumnIndexOrThrow("marca"))));
                 produto.setFornecedor(daoFornecedor.listarPorId(cursor.getString(cursor.getColumnIndexOrThrow("fornecedor"))));
 
                 produtos.add(produto);
@@ -306,7 +306,7 @@ public class DAOProduto implements DAO<Produto> {
 
                 cursorCodigoBarraFornecedor.close();
 
-                produto.setMarca(daoMarca.listarPorId(cursor.getInt(cursor.getColumnIndexOrThrow("marca"))));
+                produto.setMarca(daoMarca.listarPorId(cursor.getLong(cursor.getColumnIndexOrThrow("marca"))));
                 produto.setFornecedor(daoFornecedor.listarPorId(cursor.getString(cursor.getColumnIndexOrThrow("fornecedor"))));
 
                 produtos.add(produto);
@@ -353,7 +353,7 @@ public class DAOProduto implements DAO<Produto> {
 
                 cursorCodigoBarraFornecedor.close();
 
-                produto.setMarca(daoMarca.listarPorId(cursor.getInt(cursor.getColumnIndexOrThrow("marca"))));
+                produto.setMarca(daoMarca.listarPorId(cursor.getLong(cursor.getColumnIndexOrThrow("marca"))));
                 produto.setFornecedor(daoFornecedor.listarPorId(cursor.getString(cursor.getColumnIndexOrThrow("fornecedor"))));
             }
         }
@@ -400,7 +400,7 @@ public class DAOProduto implements DAO<Produto> {
 
                 cursorCodigoBarraFornecedor.close();
 
-                produto.setMarca(daoMarca.listarPorId(cursor.getInt(cursor.getColumnIndexOrThrow("marca"))));
+                produto.setMarca(daoMarca.listarPorId(cursor.getLong(cursor.getColumnIndexOrThrow("marca"))));
                 produto.setFornecedor(daoFornecedor.listarPorId(cursor.getString(cursor.getColumnIndexOrThrow("fornecedor"))));
             }
         }
@@ -446,7 +446,7 @@ public class DAOProduto implements DAO<Produto> {
                 cursorCodigoBarraFornecedor.close();
 
                 produto.setFornecedor(new DAOFornecedor(sqLiteDatabase).listarPorId(cursor.getString(cursor.getColumnIndexOrThrow("fornecedor"))));
-                produto.setMarca(new DAOMarca(sqLiteDatabase).listarPorId(cursor.getInt(cursor.getColumnIndexOrThrow("marca"))));
+                produto.setMarca(new DAOMarca(sqLiteDatabase).listarPorId(cursor.getLong(cursor.getColumnIndexOrThrow("marca"))));
 
                 produtos.add(produto);
             }
